propNeeded function added
I am aware that these functions don’t work perfectly as they are meant
to. After talking with Dave, I have a better understanding of how these
functions are supposed to work and what the syntax should be. I want to
learn how to properly create a package for both this exercise and to be
as prepared as possible for the midterm this Thursday. I will be coming
into office hours on Wednesday to talk with Professor Montgomery to go
over previous lectures and this exercise.